Reviews y attributes

+++ app/src/main/java/com/abel/mercadoaea/data/model/review/ResponseReview.kt
@@ -4,7 +4,7 @@ data class ResponseReview(
     val attributes: List<Any>,
     val helpful_reviews: HelpfulReviews,
     val paging: Paging,
-    val rating_average: Double,
+    val rating_average: Float,
     val rating_levels: RatingLevels,
     val reviews: List<Review>
 )
